Change the path of the `setup-foreman` action to reflect its current location in the Roblox org on GitHub (#169)
(not sure why github doesn't count this as approved)

This batch change updates the reference to the setup-foreman action to use the current location of the action in the Roblox org. This should be a no-op, or may fix currently-broken CI.
Please refer to confluence for more information.

Created by Sourcegraph batch change pdoyle/migrate-setup-foreman-action.